requirements.txt
langchain-community
langchain-ollama
langchain-core
langchain-groq
langgraph
tidb-vector
python-dotenv
streamlit
pypdf
langchain-huggingface
sentence-transformers
mysql-connector-python
